requirements.txt
scipy==1.10.1
requests==2.31.0
python-dotenv==1.0.0
shapely==2.0.1
geopandas==0.13.2